CHROMIUM: gobi: Do not resubmit int urb if read_usb submitted
Do not resubmit int urb if read_usb submitted.  To do so could lead to
race conditions in which the int urb is resubmitted while the
int_complete is processing and urb.  In the best case this causes a
message to be delayed, in the worst case, the message arrives too
late and is ignored.

BUG=chromium-os:17530
TEST=run_remote_test suite_Cellular

Change-Id: If55e7626fa7467a55d455dfb865c08465d7ad2b2
Signed-off-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3976
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>