ARM: OMAP2+: mux: warn only when GPIO really not found
omap_mux_init_gpio calls _omap_mux_init_gpio for each of the
partitions defined for the device while searching for the mux
signal requested. However, on platforms such as OMAP4 which have
multiple partitions, GPIO which is present in a later partition
will be failed to be found by _omap_mux_init_gpio in the earlier
partition search. It currently puts out a scary warning
"__omap_mux_init_gpio: Could not set gpioxx"
which is interpreted as mux configuration was faulty, which could
be:
a) a real mismatch due to typo error on behalf of the developer
or
b) just an indication that the signal was not found in an earlier
partition.

Instead move the warning to omap_mux_init_gpio to report this
warning *only after* all alternatives have been exhausted to
ensure that this warning is really for (a).

Change-Id: I243efb658aa335a82cca43232e4491c0d01c1117
Signed-off-by: Nishanth Menon <nm@ti.com>