merge_config.sh: Fix bug in final check
Arnaud Lacombe pointed out the final checking that
the requested configs were included in the final .config
was broken.

The example was that if you had a fragment that disabled
CONFIG_DECOMPRESS_GZIP applied to a normal defconfig,
there would be no final warning that CONFIG_DECOMPRESS_GZIP
was acutally set in the final .config.

This bug was introduced by me in v3 of the original patch, and
the following patch reverts the invalid change.

Reported-by: Arnaud Lacombe <lacombar@gmail.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>