i2c-s3c2410: move resume state flag to end of resume
weak memory ordering on ARM means that the register configuration
may not be complete by the time suspended is seen as 0 by another cpu,
so an i2c transaction could start with incomplete register
and gpio initialization.
careful use of memory barriers (smp_wmb() and smp_rmb())
can prevent this, but it is usually easier just to lock around
the affected code, which provides barrier semantics.

also, this sets the suspended flag and then continues with suspend.
if an i2c transaction is already in progress,
it may still be running when the suspend state is entered,
which is probably bad.

calling i2c_lock_adapter in suspend and resume would fix both problems,
preventing suspend from running while an i2c transfer is active.

Change-Id: Iab2a4f08d6a053ada5a0131e746342e5626417d8
Signed-off-by: Kisoo Yu <ksoo.yu@samsung.com>