mm: cma: Contiguous Memory Allocator added
The Contiguous Memory Allocator framework is a set of APIs for
allocating physically contiguous chunks of memory.

Various chips require contiguous blocks of memory to operate.  Those
chips include devices such as cameras, hardware video decoders and
encoders, etc.

The code is highly modular and customisable to suit the needs of
various users.  Set of regions reserved for CMA can be configured
per-platform and it is easy to add custom allocator algorithms if one
has such need.

Signed-off-by: Boojin Kim <boojin.kim@samsung.com>
Signed-off-by: Michal Nazarewicz <m.nazarewicz@samsung.com>
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>
[sanghyun75.lee@samsung.com: fixed calling of rb_root_init()]
Sigend-off-by: Sanghyun Lee <sanghyun75.lee@samsung.com>