media: vb2-ion: prevent race condition between vb2_ion_resume/suspend()
Since the time of checking to decide and making a dicision is different in
vb2_ion_resume() and vb2_ion_suspend(), they causes a problem when they
face with a race condition.

In addition to this change, vb2_ion_resume() returns an status code:
it returns positive number if it enables System MMU successfully.
the returned number is the number of request to enable System MMU.;
it returns negative number if it is failed to enable System MMU.
It never returns 0.

Change-Id: I1018c0f52bbdd8de50f036edfaa81355c1a45748
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>