iommu: exynos: chage driver according to the change in mach tree
This changes iommu driver for Exynos SoCs according to the change in
platform device definition. The change moves information about
relationship between a peripheral device and its System MMU to the
device descriptor of peripheral devices.

This patch also includes some bug fixes.

Finally,
since enabling and disabling System MMU is safe in race conditions,
no spinlock is required. But IOVMM needs to count the number of
enabling System MMU to disable it when iovmm is destroyed.

Change-Id: I0b38515bd6e9b331a1c81e235ab45afb5d06828e
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>