serial: samsung: fix device node numbering if a device is skipped
If the board file chooses not to initialize a serial port
(for example if it is handled by another driver), the
/dev/ttySACx numbers of every device after the skipped one
will change.  Modify the driver to use the platform device id
to determine the /dev/ttySACx number, which will match the
datasheet UART number (/dev/ttySAC2 will always be UART2).

Change-Id: Iedacdc395783c772834a31eccb8f0ea8b7a95000
Signed-off-by: Colin Cross <ccross@android.com>