CHROMIUM: Daisy: Change backlight PWM frequency to 1KHz
The LCD backlight on Daisy is specified to be 900-1100Hz, the original
clock for the PWM was ~1000Khz, this CL sets the PWM frequency to ~1000Hz.

This change will also impact the SMDK, it will likely resolve a similar
backight not recovering problem there.

This is roughly the same change that happened here:
  <https://gerrit-int.chromium.org/15729>
...so keeping the old commit message...

BUG=chrome-os-partner:10126
TEST=Manual, verify LCD backlight PWM is at 1000Hz.

Change-Id: I4e5c956b2d062d1ca75b36f4cb06364b40f3da42
Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>