origen: android: mali: Disable profiling of Mali L2 cache
Add patch into Origen

When connecting the ARM Streamline Performance Analyzer to Snowball
Android this will cause a board lock-up if the screen is static
(Bug #961962). The cause of this appears to be a know issue with reading
the L2 cache performance counters whilst the hardware block is powered
down.

As the current code is written to access these counters even when they
have not been explicitly selected then this issue prevents Streamline
being used reliably for any purpose.

To workaround this bug, this patch disables support for the L2 cache
counters in Mali. This has no effect on the operation of the cache.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Sangwook Lee <sangwook.lee@linaro.org>