iommu: exynos: allow multiple iovmm_setup() and one to one mapping
Calling iovmm_setup() more than once from a single device driver
will creates different address spaces for a the same device and is
error-prone behavior. It is too restricted for the device drivers
that creates device address spaces indirectly from vb2-ion because
device drivers needs to create several vb2 allocation contexts
at the same time.
This patch fixes the problem of iovmm and allows multiple call to
iovmm_setup().

In addition, this patch introduces new interfaces that create
mappings 1 to 1 mapping for the given physical area. It will be
useful for the situation that a device needs to access a buffer
through physical address while its System MMU is enabled.

Change-Id: I8eef6cf3d3879bb1f70829b64b888eb714133ca4
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>