gpu: ion: exynos: fix kernal mapping for exynos_heap
Kernel mapping from exynos_user_heap may add offset to the generated
kernel address in the kernel address of the given buffer. But the
offset is not subtracted when it is freed with vunmap().
In addition to the above problem, exynos_heap calculates the number
of pages in the buffer incorrectly.

Change-Id: I55cb9ebd2f33be184623cecc5951a19eaecfaf66
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>