HACK: ARM: exynos: switch compatibility irqs to exynos5
Commit bb19a7513dffc82f4b474ed90fc8ed691b54768b added an EXYNOS5_
prefix to the irq macros, and left compatibility #defines pointed to
the EXYNOS4_* irqs, with the expectation that the resources would be
modified at runtime on exynos5.  The compatibility irq #defines are
used in many devices that haven't been fixed for exynos5, so for now
switch the compatibility irq #defines to EXYNOS5_* if
CONFIG_ARCH_EXYNOS5 is selected and CONFIG_ARCH_EXYNOS4 is not.

Signed-off-by: Colin Cross <ccross@android.com>
Change-Id: Id50fb32dc4cc06f968250066e7691fcd1bb5b19a