ARM: b.L: secondary kernel entry code
CPUs in a big.LITTLE systems have special needs when entering the kernel
due to a hotplug event, or when resuming from a deep sleep mode.

This is vectorized so multiple CPUs can enter the kernel in parallel
without serialization.

Only the basic structure is introduced here.  This will be extended
later.

TODO: MPIDR linearization be made runtime adjusted.

Signed-off-by: Nicolas Pitre <nico@linaro.org>

Conflicts:
	arch/arm/common/Makefile