DMA: PL330: Change the cookie complete operation for cyclic mode
This patch changes the cookie complete operation on cyclic mode(DMA_CYCLIC).
It doesn't set cookies to zero because cookies is reused in case of cyclic mode.
The cookie complete operation has been changed in commit 04bf3011 (dmaengine:
provide a common function for). But, It doesn't consider cyclic mode.
This patch fixes it.

Change-Id: I3b569a14c9e42fd79c5e608f272cb31de25e9d6e
Signed-off-by: Boojin Kim <boojin.kim@samsung.com>