UPSTREAM: ARM: 7408/1: cacheflush: return error to userspace when flushing syscall fails
The cacheflush syscall can fail for two reasons:

(1) The arguments are invalid (nonsensical address range or no VMA)

(2) The region generates a translation fault on a VIPT or PIPT cache

This patch allows do_cache_op to return an error code to userspace in
the case of the above. The various coherent_user_range implementations
are modified to return 0 in the case of VIVT caches or -EFAULT in the
case of an abort on v6/v7 cores.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit c5102f5935503ebebad46e137d0eef68f272cc16)

BUG=none
TEST=none

Change-Id: Ia106e7e14d1d60ba5aa363167c9308eaf07437c7
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/19423
Reviewed-by: Doug Anderson <dianders@google.com>
Tested-by: Olof Johansson <olofj@google.com>
Commit-Ready: Doug Anderson <dianders@google.com>