CHROMIUM: Input: atmel_mxt_ts - refactor input device creation
Move input device initialization to its own helper function.
This is in preparation of a future patch that makes input device
conditional on being able to read device properties, which implies
that the device is not in its bootloader.

Just a refactor; no functional change.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:9103
TEST=builds clean; input device created as before.

Change-Id: Ife128dc63a4c23c162ed116c21cc0dd3d076a559
Reviewed-on: https://gerrit.chromium.org/gerrit/20844
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>