gpio/omap: fix dataout register overwrite and incorrect logic in runtime callback
In the existing _set_gpio_dataout_reg() implementation, the dataout
register is overwritten every time the function is called. Modify
this so that previous value is preserved until explicitly changed
by respective user of the GPIO line.

In omap_gpio_runtime_resume() the context restore should be independent
of bank->enabled_non_wakeup_gpios. This was preventing context restore
of GPIO Bank 6, dataout register. So removing the associated logic.

Signed-off-by: Tarun Kanti DebBarma <tarun.kanti@ti.com>