mfd: Add suport for MAX77686.
MAX77686 is a Mulitifunction device with PMIC, RTC and Charger on chip.
This driver provides common support for accessing the device.
This is initial version of this driver that supports to
enable the chip with its primary I2C bus. It also includes IRQ and
device tree support for MAX77686 chip.

TODO:Adding charger suport.

Signed-off-by: Yadwinder Singh Brar <yadi.brar@samsung.com>