UPSTREAM: dma: pl330: assign a new cookie when restarting tx descriptors in cyclic mode
The cookie of completed transfer descriptors are marked as zero. In case of
cyclic transfers, a new cookie needs to be assigned to the transfer
descriptors
which are picked up from the work_list list before they are re-enabled for
transfer. This prevents hitting BUG_ON in dma_cookie_complete function when
transfer descriptors are recycled from the work_list list.

Change-Id: Ib2417e61b0f69abaebd64bbb0144ff91c27eb47e
Signed-off-by: Padmavathi Venna <padma.v@samsung.com>
Signed-off-by: Thomas Abraham <thomas.abraham@linaro.org>