ARM: Exynos4: Fix the incorrect hierarchy of spi controller bus clock
The sclk_spi clock is derived currently from the first level divider
(MMCx_RATIO) which is incorrect. The output of the first level clock
is divided by a second level divider (MMCx_PRE_RATIO), the output of
which is used as the spi bus clock (sclk_spi). Fix the clock hierarchy
issues for the sclk_spi clock.

BUG=None
TEST=None

Change-Id: If64ce501d2f7386b53547603ddd7882e781d9c6f
Signed-off-by: Thomas Abraham <thomas.abraham@linaro.org>
Acked-by: Jaswinder Singh <jaswinder.singh@linaro.org>
Reviewed-on: https://gerrit-int.chromium.org/18953
Reviewed-by: Olof Johansson <olofj@google.com>
Tested-by: Doug Anderson <dianders@google.com>
Commit-Ready: Doug Anderson <dianders@google.com>