cpufreq: exynos5250: use clk_set_rate to handle frequency changes
Instead of writing to clock registers directly, use clk_set_rate
on fout_apll to change the frequency.  Allows regular clock tracing
to show the cpu frequency change steps, including reparenting the
clock onto mpll, changing the apll frequency, and reparenting the
clock back to apll.

This has two effects on the cpufreq process:
The optimization to skip switching to mpll while updating the apll
is removed.  The optimization will rarely run, as each of the 17
frequencies in the frequency table shares the same apll settings
with at most one other frequency, and its optimizing out what is
already a very short time (50 us in my tests).

It also moves the reparenting of the cpu to mpll outside of the
calls to reprogram the clock dividers.  This means that at low
frequencies, the frequency of the clocks divided off the cpu
will jump up to the current divider applied to 800 MHz, but this
could already happen because the clock divider values were not
updated to take into account mpll when moving from a high
frequency to a low frequency.

Change-Id: I194f38d75961d94dd928944192bff0cb0d8b4ebe
Signed-off-by: Colin Cross <ccross@android.com>