drm/exynos: Adding IOMMU mapping to NON-CONTIG gem framework
The non-contig gem buffers are only used for CPU access currently.
This patch adds IOMMU support for these gem buffers. When a gem buffer
is created with NONCONTIG flag, a SGT is created with the pages
allocated. This patch maps the SGT to create IOMMU mapping by calling
dma_map_sg.

Change-Id: I3df514254c4be714908d13fc9c6a5967e9a07df0
Signed-off-by: Prathyush K <prathyush.k@samsung.com>