ARM: SAMSUNG: add support platfrom_data for prescaler on PWM
There are 2 prescalers, prescaler0 and prescaler1 in SAMSUNG PWM
hardware and each prescaler is shared on 3 and 2 PWMs. Now we need
to initialize the prescaler values for PWM and it should be called
in pwm_probe() and pwm_resume() of PWM driver. This patch adds the
platform_data for prescalers of PWM Timer.

Change-Id: I09631fedb059248a45c1acb6199accc42d1250b2
Signed-off-by: Kukjin Kim <kgene.kim@samsung.com>