CHROMIUM: exynos: Ensure the manual update bit is off before setting it.
When updating the tcntb and tcmpb registers of one of the exynos5's PWMs, you
have to turn the manual update bit of the tcon register on (and then ideally
back off) to make the hardware update the underlying tcnt and tcmp registers
which are actually driving the PWM output. If the manual update bit is already
set, setting it again or turning it off doesn't do anything. This change makes
the exynos driver more resilient by explicitly turn off that bit before turning
it on to install a configuration.

This change is necessary currently because something during initialization is
turning on that bit and leaving it on, and that makes configuration of the
backlight fail. I believe this problem was masked before because the PWM
driving the backlight used to be disabled when going from U-Boot to the kernel.
Disabling and then re-enabling the PWM has the same effect as the manual update
bit.

We should identify what is leaving the manual update bit on and make it stop,
but even after that problem is fixed, it wouldn't be a bad idea to leave this
change in place since it protects against bugs like this in the future.

BUG=chrome-os-partner:10126
TEST=Built and booted on Daisy with the backlight PWM enabled in U-Boot. Saw
the backlight continue to operate once the system was booted where before it
would turn off and not respond to its PWM input.

Change-Id: If012696ae11fb5d42a6d8982211afd18fcebc7e7
Signed-off-by: Gabe Black <gabeblack@google.com>
(cherry picked from commit 85b15bc9409da33b8d2c64e2e34957ef2f23cb8a)