CHROMIUM: drm/i915: Replace pin count BUG_ON with warning
This patch reduces the BUG_ON when pin_count reaches max to a warning.
The reason we want to just warn is because the unpin_work_fn work
function can get starved out during soft lockup. When the unpin worker
is starved, the pin_count rises, and the gem BUG_ON races the soft
lockup crash reporter.

The original intent of this BUG was to catch overflow cases when all of
the small bitfields in drm_i915_gem_object were combined into one
bitfield (commit 778c35444). Since pin leaks aren't an issue in steady
state, it should be fine to downgrade the error.

Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=chrome-os-partner:8552,chromium-os:30733
TEST=Tested on lumpy

Change-Id: Ibd7377385f36a722b588cfe4989b3e26d2647bd2
Reviewed-on: https://gerrit.chromium.org/gerrit/22829
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Ready: Sean Paul <seanpaul@chromium.org>