UPSTREAM: serial: samsung: Fixed wrong comparison for baudclk_rate
port->baudclk_rate should be compared to the rate of port->baudclk,
because port->baudclk_rate was assigned as the rate of port->baudclk previously.
So to check that the current baudclk rate is same as previous rate,
the target of comparison sholud be the rate of port->baudclk.

BUG=None
TEST=Built and booted on one board.

Signed-off-by: Naveen Krishna Chatradhi <naveen@chromium.org>
(Landed on linux-next on Sun May 20, cherry picked from commit
25f04ad423e5eb40c33a904db5a0d2c7e3bf08f5)

Change-Id: Iae21f911f63ad483289fbd85a3123a609e7aa50d
Signed-off-by: Jun-Ho, Yoon <junho78.yoon@samsung.com>
Signed-off-by: Kyoungil Kim <ki0351.kim@samsung.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33419
Tested-by: naveen krishna chatradhi <naveen@chromium.org>
Reviewed-by: Arjun.K.V <arjun.kv@samsung.com>
Commit-Ready: Doug Anderson <dianders@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>