CHROMIUM: exynos: Stop the PWM timer while configuring it.
The Samsung PWM platform driver configures the PWM it's associated with, but
it leaves it running while it does so. There's no mention of this in the
manual, but it seems to be the case that you can't manually set the count or
compare registers while the timer is running. This change adds some code which
checks whether the PWM is running, stops it, does the configuration, and if
the PWM was running originally turns it back on.

BUG=chrome-os-partner:10126
TEST=Booted on Daisy and saw the backlight work in both U-Boot and ChromeOS.

Change-Id: Iabdfaff69a95241bc4723722448a895b8f860926
Signed-off-by: Gabe Black <gabeblack@google.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 994ae76b1454138ea70dba2c1d2f94f5aa307796)
(dianders: slight cleanup with converting bitfield to bool)
Reviewed-on: https://gerrit-int.chromium.org/18779
Commit-Ready: Doug Anderson <dianders@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>
Tested-by: Doug Anderson <dianders@google.com>