CHROMIUM: exynos: dts: Take "smdk5250" out of daisy/snow dts
Daisy and snow have some important differences from smdk5250.  Taking
this out allows us to check for smdk5250.  Note that we really
shouldn't be testing for the board itself in the long term, but in the
short term it is helpful.

BUG=None
TEST=Compile and boot; notice that future patch with test for smdk5250
detects that snow is not smdk5250.

Change-Id: I52f01cc7355719e46dcd4dcf09ec9affa399133d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/18459
Tested-by: Doug Anderson <dianders@google.com>
Reviewed-by: Abhilash Kesavan <a.kesavan@samsung.com>
Commit-Ready: Doug Anderson <dianders@google.com>