regulator: tps65090: move to device tree based probing
Move away from the platform device probing to a model that only uses device-tree.

There were no in-tree users of the old configuration method, and all known platforms
where this device is in use are ARM platforms that are enabled for device tree,
which makes this a good target for this conversion.

BUG=chrome-os-partner:9776
TEST=build, boot, see graphics on console

Change-Id: I1fc85512ba5757ae7ec598781ea3da5d4adacace
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/18607
Reviewed-by: Doug Anderson <dianders@google.com>
Commit-Ready: Olof Johansson <olofj@google.com>
Tested-by: Olof Johansson <olofj@google.com>