media: tv: Switch to i2c8 for HDMI-PHY on Exynos
Exynos5 has 8 i2c channels and a dedicated internal one for the
HDMI-PHY. We setup the HDMI-PHY with an i2c8 alias as it would
remain constant across boards irrespective of usage of other channels.
Modify the hdmi_driver_data to select the HDMI-PHY bus as 8 to match
this.

BUG=None
TEST=Only build tested, no h/w at the moment; The HDMI-PHY initialization
errors should go away.

Change-Id: I3c7482d83a3641902cabf091d94a925d666db9ec
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/18568
Reviewed-by: Doug Anderson <dianders@google.com>
Tested-by: Doug Anderson <dianders@google.com>
Commit-Ready: Olof Johansson <olofj@google.com>