mfd: tps65090: remove platform data
Platform data isn't really needed for this MFD driver; the regulator
data is not passed in through it and the MFD subdev information isn't
really a function of board configuration. So it makes perfect sense to
just hardcode the cells in the driver and not having to specify them on
every board, etc.

BUG=chrome-os-partner:9776
TEST=build, boot, see graphics on console

Change-Id: I94ddd28b037027fdfab5d41a741d8ffedac97ae8
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/18606
Reviewed-by: Doug Anderson <dianders@google.com>
Tested-by: Olof Johansson <olofj@google.com>