hwspinlock: modify API to select mutex or spinlock
In the current implementation of hwspinlock driver, if a client
wishes to use it's locking APIs from a context that may sleep it
results in a kernel warning backtrace since it uses spinlock
for locking within the driver. This patch modifies the
hwspin_lock_request API which passes an extra parameter to let
the client choose which lock the driver may use.

If a client wishes to call the hwspinlock APIs only from a context
that may sleep it can pass USE_MUTEX_LOCK parameter but if it wishes
to call it in interrupt context or any other context it can pass
USE_SPIN_LOCK parameter. In this case it will not be able to call
these APIs from a context that sleeps.

This was required to accomodate clients like I2C driver which
call these APIs from a context that may sleep.

Change-Id: Iefc0b266fa3ed702aad425d6000e4709c768a40c
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>