CHROMIUM: exynos: pwm: Cosmetic tidy of PWM variable names
This is based on a similar change by Gabe Black in U-Boot titled
"exynos: Tidy up the pwm_config function in the exynos pwm driver".
It fixes a few things, specifically:
* The define NS_IN_HZ doesn't make sense.  This is the number of
  nanoseconds in a second, not in a hertz.
* The variable "period" was used to store a frequency, so name it
  "frequency" instead of period.

Gabe's original U-Boot change also had some functional changes in it
to try to avoid a special case.  We can revisit whether that's a good
idea in a separate change.  This change itself has no functional
changes.

BUG=None
TEST=Compiled; tested that backlight still works OK.

Change-Id: I800e9837e6df0a0a7a17eb5dc5fac99752cb0cbf
Signed-off-by: Doug Anderson <dianders@chromium.org>