EXYNOS5: GPU: Makefile hack for vithar
This is ugly, can be pruned, and should be replaced with an include file
that contains all the base defines. Heck, it should be replaced with an
include file instead. Or runtime config. Or something.

Actually, the whole driver needs rewriting/cleanup, so I just did the
bare minimum. Only 15 pieces of flair for me today.

Change-Id: I1204a5298bc33718706b46b13b94705f6fcaf3a5
Signed-off-by: Olof Johansson <olofj@chromium.org>