drm/exynos: Add IOMMU support to DRM and DRM-FIMD
DRM and its subdevices like DRM-FIMD need to have a common
IOMMU mapping which is attached to all its devices.

An IOMMU mapping is first created in DRM-FIMD probe. The driver has to find
a 'sysmmu' field in its of_node and retrieves the corresponding sysmmu's
platform device. A IOMMU mapping is then created.

All the other devices will use the same mapping which is stored as
'exynos_drm_common_mapping'.

IOMMU mapping is also added to DRM probe. DRM is a virtual device which
exists as a platform device in the mach file. It uses the same mapping
which was created by DRM-FIMD. All the allocations/mapping go through
DRM device which will create a page table entry for each frame buffer.
Since the page table base is same for DRM/DRM-FIMD/DRM-HDMI (since IOMMU
mapping is shared), DRM-FIMD/DRM-HDMI can use the same dma-address
and access the buffer.

Change-Id: I91d18984b9fda0883361b6abf7e95f67ebe4ad9e
Signed-off-by: Prathyush K <prathyush.k@samsung.com>