ARM: EXYNOS: SYSMMU: Add IOMMU machine support to Exynos5
For creating IOMMU mapping, the dts file is modified to add a
'sysmmu' field for all the parent devices e.g. MFC, GSC, HDMI etc.
When the parent device is getting probed, the corresponding sysmmu
device is retrieved and a IOMMU mapping is created.

This patch adds two functions to dev-sysmmu which are called during
probe of respective parent devices like MFC/FIMD etc.
1> s5p_create_iommu_mapping is modified to get a mapping as input. This
is used for sharing a IOMMU mapping across multiple devices. A new
mapping is created only if the received mapping is NULL and the function
also returns this mapping.
2> find_sysmmu_dt is a common function which retrieves the sysmmu platform
device from the parent device's of_node.

Change-Id: Ic0e324fdaa758229f6d2ef8e00175346230716b4
Signed-off-by: Prathyush K <prathyush.k@samsung.com>