CHROMIUM: chromeos-ec: use level-triggered interrupt
The EC made this change a while ago to better support the firmware,
which checks for keyboard changes by polling the gpio line which
the kernel uses for interrupts.

With this change, there is still an issue in the kernel.
We see two interrupts for each keyboard state change.
One theory is that the level-triggered interrupt generic code
does not clear and unmask the interrupt in the correct order.
Functionally this doesn't cause problems so we'll fix it separately.

There may be a spew of interrupts between probing of chromeos-ec
and probing of mkbp.  However, this hasn't been observed.

Also, improve logging.

BUG=chrome-os-partner:10172
TEST=tested manually, verified that bug is gone and kbd still works

Change-Id: I5f244b82c480b154e6744d496bcf800f141906f4
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/19093
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Ready: Luigi Semenzato <semenzato@google.com>
Tested-by: Luigi Semenzato <semenzato@google.com>
Reviewed-by: Doug Anderson <dianders@google.com>